notes-app/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9